additions to abs functions

+++ include/libft.h
@@ -6,7 +6,7 @@
 /*   By: tpolonen <tpolonen@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/13 14:07:36 by tpolonen          #+#    #+#             */
-/*   Updated: 2022/06/06 15:42:52 by tpolonen         ###   ########.fr       */
+/*   Updated: 2022/06/06 17:56:10 by tpolonen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,46 +35,46 @@ typedef struct s_dstr
  * Libc functions
  */
 
-void	*ft_calloc(size_t nmemb, size_t size);
-void	*ft_memset(void *b, int c, size_t len);
-void	ft_bzero(void *s, size_t n);
-void	*ft_memcpy(void *dst, const void *src, size_t n);
-void	*ft_memccpy(void *dst, const void *src, int c, size_t n);
-void	*ft_memmove(void *dst, const void *src, size_t len);
-void	*ft_memchr(const void *s, int c, size_t n);
-void	*ft_memrchr(const void *s, int c, size_t n);
-int		ft_memcmp(const void *s1, const void *s2, size_t n);
-size_t	ft_strlen(char const *str);
-char	*ft_strdup(const char *str);
-char	*ft_strcpy(char *dst, const char *src);
-char	*ft_strncpy(char *dst, const char *src, size_t len);
-char	*ft_strcat(char *s1, const char *s2);
-char	*ft_strncat(char *s1, const char *s2, size_t n);
-size_t	ft_strlcat(char *dst, const char *src, size_t dstsize);
-size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize);
-char	*ft_strchr(const char *s, int c);
-char	*ft_strrchr(const char *s, int c);
-char	*ft_strchrnul(const char *s, int c);
-char	*ft_strstr(const char *haystack, const char *needle);
-long	ft_strtol(const char *nptr, char **endptr);
-char	*ft_strnstr(const char *haystack, const char *needle, size_t len);
-int		ft_strcmp(const char *s1, const char *s2);
-int		ft_strncmp(const char *s1, const char *s2, size_t n);
-int		ft_atoi(const char *str);
-int		ft_abs(int n);
-int		ft_llabs(long long n);
-int 	ft_ssabs(ssize_t n);
-int		ft_max(int a, int b);
-int		ft_isalpha(int c);
-int		ft_isdigit(int c);
-int		ft_isalnum(int c);
-int		ft_isascii(int c);
-int		ft_isspace(int c);
-int		ft_isprint(int c);
-int		ft_isupper(int c);
-int		ft_islower(int c);
-int		ft_toupper(int c);
-int		ft_tolower(int c);
+void		*ft_calloc(size_t nmemb, size_t size);
+void		*ft_memset(void *b, int c, size_t len);
+void		ft_bzero(void *s, size_t n);
+void		*ft_memcpy(void *dst, const void *src, size_t n);
+void		*ft_memccpy(void *dst, const void *src, int c, size_t n);
+void		*ft_memmove(void *dst, const void *src, size_t len);
+void		*ft_memchr(const void *s, int c, size_t n);
+void		*ft_memrchr(const void *s, int c, size_t n);
+int			ft_memcmp(const void *s1, const void *s2, size_t n);
+size_t		ft_strlen(char const *str);
+char		*ft_strdup(const char *str);
+char		*ft_strcpy(char *dst, const char *src);
+char		*ft_strncpy(char *dst, const char *src, size_t len);
+char		*ft_strcat(char *s1, const char *s2);
+char		*ft_strncat(char *s1, const char *s2, size_t n);
+size_t		ft_strlcat(char *dst, const char *src, size_t dstsize);
+size_t		ft_strlcpy(char *dst, const char *src, size_t dstsize);
+char		*ft_strchr(const char *s, int c);
+char		*ft_strrchr(const char *s, int c);
+char		*ft_strchrnul(const char *s, int c);
+char		*ft_strstr(const char *haystack, const char *needle);
+long		ft_strtol(const char *nptr, char **endptr);
+char		*ft_strnstr(const char *haystack, const char *needle, size_t len);
+int			ft_strcmp(const char *s1, const char *s2);
+int			ft_strncmp(const char *s1, const char *s2, size_t n);
+int			ft_atoi(const char *str);
+int			ft_abs(int n);
+long long	ft_llabs(long long n);
+size_t 		ft_ssabs(ssize_t n);
+int			ft_max(int a, int b);
+int			ft_isalpha(int c);
+int			ft_isdigit(int c);
+int			ft_isalnum(int c);
+int			ft_isascii(int c);
+int			ft_isspace(int c);
+int			ft_isprint(int c);
+int			ft_isupper(int c);
+int			ft_islower(int c);
+int			ft_toupper(int c);
+int			ft_tolower(int c);
 
 /*
  * Additional functions
